fix: `TeamMembersTable` column.id issue (#70)
issue: #70
